+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/Exceptions/AddressablesSupportDisabledException.cs
@@ -5,16 +5,14 @@ using JetBrains.Annotations;
 namespace Eflatun.SceneReference.Exceptions
 {
     /// <summary>
-    /// Thrown if an operation that requires addressables support is attempted while addressables support is disabled.<p/>
-    /// You can avoid this exception by making sure addressables support is enabled.
+    /// Thrown if an operation that requires addressables support is attempted while addressables support is disabled.
+    /// To fix it, make sure addressables support is enabled.
     /// </summary>
     [PublicAPI]
     [Serializable]
     public class AddressablesSupportDisabledException : SceneReferenceException
     {
-        private const string ExceptionMessage =
-            "An operation that requires addressables support is attempted while addressables support is disabled." +
-            "\nYou can avoid this exception by making sure addressables support is enabled.";
+        private const string ExceptionMessage = "An operation that requires addressables support is attempted while addressables support is disabled. To fix it, make sure addressables support is enabled.";
 
         internal AddressablesSupportDisabledException() : base(ExceptionMessage)
         {